+++ types.go
@@ -27,20 +27,20 @@ type Ground struct {
 // Temple is a special place in the map with fixed cost to pass through it,
 // Code irrelevante -> the cost are little bite higher than the ground because the temple keep the Gold Knights.
 type Temple struct {
-	Name       string `json:"name"`
-	Difficulty int    `json:"difficulty"`
-	Position   Point  `json:"position"`
+	Name       string  `json:"name"`
+	Difficulty float64 `json:"difficulty"`
+	Position   Point   `json:"position"`
 }
 
 type Saint struct {
 	Name  string  `json:"name"`
-	Power float32 `json:"power"`
+	Power float64 `json:"power"`
 	Lives int     `json:"lives"`
 }
 
 //Environment of the game
 type Environment struct {
-	AvailableTime int        `json:"availableTime"`
+	AvailableTime float64    `json:"availableTime"`
 	Start         Point      `json:"start"`
 	End           Point      `json:"end"`
 	Grounds       []Ground   `json:"grounds"`
